Consolidate health sidebar tests
